ubdsrv: fix allocation bug
struct ubdsrv_queue {
        int q_id;
        struct ubd_io ios[0];
}

If the struct contains a  zero-length array,
we should not get queue via dev->queues[i] directly
because the size of struct is not equal to the real size
(with some ubd_io structs).

Therefore we change dev->queues to an array(fixed size) of pointers.
Then each struct ubdsrv_queue is allocated with its real size.

Signed-off-by: Ziyang Zhang <ZiyangZhang@linux.alibaba.com>